# backend/src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: The Fellowship of the Code Flood checker
  description: A flood checker that can be installed in municipalities for proactive work against flooding.
  version: 1.0.0
servers:
  - url: http://localhost:5001/
    description: "This is the current testing and production enviroment for this API"
paths:
...
  /users:
    get:
      tags:
        - User observation
      summary: Get all user observations with optional sorting
      description: |
        Returns all user-submitted observations, optionally sorted by a valid field.
        If the `sorting` parameter is not provided or invalid, the default sort is by `id`.

        Allowed sort fields:
          - id
          - location
          - timestamp
          - riskAssesment
          - waterlevel
      parameters:
        - in: query
          name: sorting
          required: false
          schema:
            type: string
            enum:
              - id
              - location
              - timestamp
              - riskAssesment
              - waterlevel
          description: Field to sort results by (default is `id`)
      responses:
        "200":
          description: Observations successfully retrieved and sorted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Observations sorted by timestamp
                  user_observations:
                    type: array
                    items:
                      $ref: "#/components/schemas/user_observation"
        "404":
          description: No observations found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: No observations found.
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error.
...
  /users/tips/postTip:
    post:
      tags:
        - tips
      summary: Creates a tip
      description: Creates a tip and adds it to the tips array
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userTipsObject"
      responses:
        "200":
          description: Tip created!
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tip:
                    $ref: "#/components/schemas/userTipsObject"
  /users/tips/putTip/{id}:
    put:
      tags:
        - tips
      summary: Edits a tip
      description: Updates an existing tip by its ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the tip to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userTipsObject"
      responses:
        "200":
          description: Tip updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tips:
                    type: array
                    items:
                      $ref: "#/components/schemas/userTipsObject"
  /users/tips/deleteTip/{id}:
    delete:
      tags:
        - tips
      summary: Delete a tip
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the tip to delete
      responses:
        "200":
          description: Tip deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tipsToDelete:
                    type: array
                    items:
                      $ref: "#/components/schemas/userTipsObject"
        "404":
          description: Tip not found
        "500":
          description: Internal server error
...
components:
  schemas:
    userTipsObject:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        location:
          type: string
        description:
          type: string
    user_observation:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        location:
          type: string
        warning:
          type: string
        waterlevel:
          type: string
        risk_assesment:
          type: string
        description:
          type: string
        proactive_actions:
          type: string
...